First pass of callback-based user list
Insert and remove entries from the user list to handle most common
operations. (JOIN, PART, QUIT, NICK, MODE) The whole list still gets
clobbered and re-rendered for more drastic things like switching between
channels. When a user is added or removed from the sorted list of users,
callbacks are fired, which update the DOM to reflect changes in the list.